+++ Mealscan/ViewController.m
@@ -31,7 +31,7 @@
   dispatch_async(dispatch_get_main_queue(), ^(void){
     NSDataAsset * frozen = [[NSDataAsset alloc] initWithName:@"FrozenPB"];
     char display[1024] = "";
-    greetings(display, frozen.data.bytes, frozen.data.length);
+    greetings(display, frozen.data.bytes, (int)frozen.data.length);
     self->_textStatus.text = [NSString stringWithUTF8String:display] ;
   });
 }
